requirements.txt
beautifulsoup4==4.7.0
certifi==2018.11.29
chardet==3.0.4
cycler==0.10.0
fake-useragent==0.1.11
idna==2.8
kiwisolver==1.1.0
lxml==4.2.5
matplotlib==3.0.3
numpy==1.15.4
pandas==0.23.4
Pillow==6.0.0
pyparsing==2.4.0
python-dateutil==2.7.5
pytz==2018.7
requests==2.21.0
selenium==3.141.0
six==1.12.0
soupsieve==1.6
urllib3==1.24.1
